ECO-2237: Frontend: compatibility with shop suite
