+++ src/SprykerEco/Yves/Braintree/Theme/default/components/molecules/credit-card/credit-card.twig
@@ -1,54 +1,42 @@
-{% extends model('component') %}
+{% extends molecule('payment-form', 'Braintree') %}
 
 {% define config = {
     name: 'credit-card',
     tag: 'credit-card'
 } %}
 
-{% define data = {
-    form: required,
-    scriptSrc: 'https://js.braintreegateway.com/js/braintree-2.32.1.min.js'
-} %}
-
-{% define attributes = {
-    'data-braintree-client-token': data.form.braintreeCreditCard.vars.clientToken
-} %}
-
 {% block body %}
-    {% include molecule('script-loader') with {
-        attributes: {
-            src: data.scriptSrc
-        }
-    } only %}
+    {{ parent() }}
 
-    <div class="payment-method payment-subform braintree-form braintree-credit-card-form payment-form small-12 medium-9 large-6" id="braintreeCreditCard">
-        <div class="braintree-method braintree-credit-card-method">
-            <div class="small-12 columns">
-                <label for="braintree-credit-card-number">
+    <div class="grid" id="braintreeCreditCard">
+        <div class="col col--sm-12 form__fields {{ config.jsName }}__method">
+            <div class="form__field">
+                <label for="{{ config.name }}__number">
                     {{ 'braintreeCreditCard.credit_card_number' | trans }}*
                 </label>
-                <div id="braintree-credit-card-number" class="hosted-field"></div>
+                <div id="{{ config.name }}__number" class="hosted-field {{ config.jsName }}__number"></div>
             </div>
 
-            <div class="small-12 columns">
-                <label for="braintree-credit-card-cvv">
+            <div class="form__field">
+                <label for="{{ config.name }}__cvv">
                     {{ 'braintreeCreditCard.ccv' | trans }}*
                 </label>
-                <div id="braintree-credit-card-cvv" class="hosted-field"></div>
+                <div id="{{ config.name }}__cvv" class="hosted-field {{ config.jsName }}__cvv"></div>
             </div>
 
-            <div class="small-12 columns">
-                <label for="braintree-credit-card-expiration-date">
+            <div class="form__field">
+                <label for="{{ config.name }}__expiration-date">
                     {{ 'braintreeCreditCard.expiration_date' | trans }}*
                 </label>
-                <div id="braintree-credit-card-expiration-date" class="hosted-field"></div>
+                <div id="{{ config.name }}__expiration-date" class="hosted-field {{ config.jsName }}__expiration-date"></div>
             </div>
 
-            <div class="braintree-error braintree-credit-card-error"></div>
+            <div class="form__field">
+                <div class="{{ config.jsName }}__error"></div>
+            </div>
 
             {{ form_errors(data.form.braintreeCreditCard) }}
             {{ form_widget(data.form.braintreeCreditCard) }}
         </div>
     </div>
-
 {% endblock %}
