+++ src/SprykerEco/Yves/Braintree/Theme/default/components/molecules/pay-pal/index.ts
@@ -1,4 +1,3 @@
-import './credit-card.scss';
 import register from 'ShopUi/app/registry';
-export default register('credit-card', () => import(/* webpackMode: "lazy" */'./credit-card'));
+export default register('pay-pal', () => import(/* webpackMode: "lazy" */'./pay-pal'));
