+++ src/SprykerEco/Yves/Braintree/Theme/default/components/molecules/payment-form/payment-form.twig
@@ -1,8 +1,8 @@
 {% extends model('component') %}
 
 {% define config = {
-    name: 'credit-card',
-    tag: 'credit-card'
+    name: 'payment-form',
+    tag: 'payment-form'
 } %}
 
 {% define data = {
@@ -20,37 +20,4 @@
             src: data.scriptSrc
         }
     } only %}
-
-    <div class="grid" id="braintreeCreditCard">
-        <div class="col col--sm-12 form__fields {{ config.jsName }}__method">
-            <div class="form__field">
-                <label for="{{ config.name }}__number">
-                    {{ 'braintreeCreditCard.credit_card_number' | trans }}*
-                </label>
-                <div id="{{ config.name }}__number" class="hosted-field {{ config.jsName }}__number"></div>
-            </div>
-
-            <div class="form__field">
-                <label for="{{ config.name }}__cvv">
-                    {{ 'braintreeCreditCard.ccv' | trans }}*
-                </label>
-                <div id="{{ config.name }}__cvv" class="hosted-field {{ config.jsName }}__cvv"></div>
-            </div>
-
-            <div class="form__field">
-                <label for="{{ config.name }}__expiration-date">
-                    {{ 'braintreeCreditCard.expiration_date' | trans }}*
-                </label>
-                <div id="{{ config.name }}__expiration-date" class="hosted-field {{ config.jsName }}__expiration-date"></div>
-            </div>
-
-            <div class="form__field">
-                <div class="{{ config.jsName }}__error"></div>
-            </div>
-
-            {{ form_errors(data.form.braintreeCreditCard) }}
-            {{ form_widget(data.form.braintreeCreditCard) }}
-        </div>
-    </div>
-
 {% endblock %}
